[GH-123] moved confirmation menu
